+++ R/mw_ui.R
@@ -18,8 +18,8 @@
 #' @noRd
 mwUI <- function(ns, inputs, nrow = 1, ncol = 1, outputFun = NULL,
                  okBtn = TRUE, saveBtn = TRUE, updateBtn = FALSE,
-                 areaBtns = TRUE, border = FALSE,
-                 width = "100%", height = "400px", fillPage = TRUE) {
+                 areaBtns = TRUE, border = FALSE, width = "100%", height = "400px",
+                 fillPage = TRUE, showCompare = TRUE) {
 
   htmldep <- htmltools::htmlDependency(
     "manipulateWidget",
@@ -40,7 +40,7 @@ mwUI <- function(ns, inputs, nrow = 1, ncol = 1, outputFun = NULL,
         style = paste("width:", width, ";height:", height, ";"),
         fillRow(
           flex = c(NA, NA, 1),
-          .uiMenu(ns, inputs$ncharts, nrow, ncol, showSettings, okBtn, saveBtn, updateBtn, areaBtns),
+          .uiMenu(ns, inputs$ncharts, nrow, ncol, showSettings, okBtn, saveBtn, updateBtn, areaBtns, showCompare),
           .uiInputs(ns, inputs),
           .uiChartarea(ns, inputs$ncharts, nrow, ncol, outputFun)
         )
@@ -95,7 +95,7 @@ mwUI <- function(ns, inputs, nrow = 1, ncol = 1, outputFun = NULL,
   )
 }
 
-.uiMenu <- function(ns, ncharts, nrow, ncol, settingsBtn, okBtn, saveBtn, updateBtn, areaBtns) {
+.uiMenu <- function(ns, ncharts, nrow, ncol, settingsBtn, okBtn, saveBtn, updateBtn, areaBtns, showCompare = TRUE) {
   container <- tags$div(
     class="mw-menu"
   )
@@ -113,7 +113,7 @@ mwUI <- function(ns, inputs, nrow = 1, ncol = 1, outputFun = NULL,
     container <- tagAppendChild(container, settingsBtn)
   }
 
-  if (areaBtns && ncharts > 1) {
+  if ((areaBtns && ncharts > 1) &  showCompare){
     container <- tagAppendChild(container, .uiChartBtns(ns, ncharts, nrow, ncol))
   }